move test against date command to integration tests
